Tank Movement and ReMapping controls feature
Tanks can now move to highlighted sqares. You can also reassign the controls for different actions. As default I made Move the M button and Turn the T button. To move the tank, you press M then click on the tank you want to move, then click on the square to move it to.